Change the default genesis `max_transaction_cpu_usage` from 150ms to 475ms to allow for slow ci/cd. All the integration tests already used 475ms for `max_transaction_cpu_usage`. This PR changes the performance tests to use the same value. Also changed `max_block_cpu_usage` from 200ms to 500ms since it has to be larger than `max_transaction_cpu_usage`. This also matches the value used for integration tests.

Note #1401 was not sufficient because the value in the performance `genesis.json` was being overwritten by `launcher.py`.

Resolves #1416